% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsbp.R
\name{getPart}
\alias{getPart}
\title{Retrieve information for a part from the Registry of Standard Biological Parts}
\usage{
getPart(id)
}
\arguments{
\item{id}{A part name.}
}
\value{
A data frame containing part information.
}
\description{
Retrieve information for a part from the Registry of Standard Biological Parts
}
\section{Information Guidelines}{

The provided information follows the guidelines set forth by iGEM Foundation at http://parts.igem.org/Registry_API/Guidelines.
}

\section{Content License}{

All Registry content falls under Creative Commons Attribution-ShareAlike: https://creativecommons.org/licenses/by-sa/4.0/.
}

\examples{
getPart("BBa_R0040")
}
